templates/: Add base.html and use it (#108)
* templates: Add base.html and use it

This commit adds a base.html containing
the common html code to all the time. This
file can be extended in other templates,
and can avoid to re-write the header and
footer again 'n' again.

Closes https://github.com/Utkarsh1308/TabOverSpace/issues/94, https://github.com/Utkarsh1308/TabOverSpace/issues/70

* home/: Move assets out of home sub-dir

The assets are common to all django-apps,
so instead of keeping it in home sub-dir
in the statics directory, keep it as a
sub-dir under static dir. Not a sub-dir
of home sub-dir i.e. "/home/static/home/assets/"
should be changed to "/home/static/assets/"