Fix Database Issue (#132)
https://github.com/Utkarsh1308/TabOverSpace/pull/124 modified the
Question model and changed one of it's arguments from question to title so
we were not getting back any results when calling the x.question since it doesn
not exist resulting in NoReverseMatch error.

This PR fixes that issue. For now there are no questions present in the database. 
We need to add the questions in a future PR.

Closes https://github.com/Utkarsh1308/TabOverSpace/issues/131